Add files via upload
This was some Python practice using Jupyter Notebooks. First I imported a few libraries, Numpy, and Matplotlib. Numpy is one of the most popular Python libraries for array manipulation and numeric computing. Matplotlib is the most popular visualization library in the Python ecosystem. After that, I generated some plots.

Then I imported two libraries, requests to pull data from the web, and pandas to process it. After that, I used a predefined function that simplified the process of importing data from Cryptowatch. I pulled the data from Bitcoin and Ether, two of the most popular cryptocurrencies, for the last 7 days. I was able to pull data from the internet with a few lines of code and create a DataFrame and plot it all within Jupyter Lab. I used the plotting library Bokeh to generate interactive plots which can be manipulated right within the browser. lastly, I generated an Excel file from the downloaded prices